horizonBackend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.6
django-db-connection-pool==1.2.5
django-environ==0.12.0
djangorestframework==3.15.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
SQLAlchemy==2.0.38
sqlparams==6.2.0
sqlparse==0.5.3
typing_extensions==4.12.2
